moving testing class to own .py file
